+++ packages/blend/lib/components/CodeV2/CodeEditorV2/CodeEditorV2.tsx
@@ -1,9 +1,9 @@
 import { forwardRef, useState } from 'react'
 import Block from '../../Primitives/Block/Block'
 import { useResponsiveTokens } from '../../../hooks/useResponsiveTokens'
-import type { CodeV2Tokens } from '../tokens'
-import { CodeV2Variant } from '../types'
-import { CodeEditorV2Props } from './codeEditorV2.types'
+import type { CodeV2Tokens } from '../codeV2.tokens'
+import { CodeV2Variant } from '../codeV2.types'
+import { CodeEditorV2Props } from '../codeV2.types'
 import { createCopyToClipboard } from '../utils'
 import { shouldShowLineNumbers, getContainerStyles } from '../utils'
 import { CodeV2Header } from '../CodeV2Header'
