+++ packages/blend/lib/components/CodeV2/MonacoEditor/MonacoEditorWrapper.tsx
@@ -2,9 +2,9 @@ import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import Editor, { OnMount } from '@monaco-editor/react'
 import type * as Monaco from 'monaco-editor'
 import Block from '../../Primitives/Block/Block'
-import type { CodeV2Tokens } from '../tokens'
+import type { CodeV2Tokens } from '../codeV2.tokens'
 import { createEditorTheme } from './monacoTheme'
-import { MonacoEditorWrapperProps } from '../types'
+import { MonacoEditorWrapperProps } from '../codeV2.types'
 import {
     BLEND_EDITOR_THEME_NAME,
     EDITOR_FOCUS_DELAY_MS,
