+++ packages/blend/lib/components/CodeV2/codeV2.tokens.ts
@@ -1,9 +1,9 @@
 import type { CSSObject } from 'styled-components'
-import { DiffLineType } from './types'
+import { DiffLineType } from './codeV2.types'
 import { BreakpointType } from '../../breakpoints/breakPoints'
 import { FoundationTokenType } from '../../tokens/theme.token'
-import { darkCodeTokens } from './code.dark.tokens'
-import { lightCodeTokens } from './code.light.token'
+import { darkCodeTokens } from './codeV2.dark.tokens'
+import { lightCodeTokens } from './codeV2.light.token'
 import { Theme } from '../../context/theme.enum'
 
 export type CodeV2Tokens = {
